services/numa-api/requirements.txt
# Core Framework
fastapi
uvicorn[standard]
python-multipart

# Database & ORM
sqlalchemy
pydantic
pydantic[email]

# Utilities
python-dotenv

# Security & Authentication
python-jose[cryptography]
passlib[bcrypt]

# AI & Machine Learning
openai-whisper

# Testing
pytest
pytest-asyncio
httpx

# Code formatting
black
isort
